src/ingestion/requirements.txt
functions-framework
cloudevents
google-cloud-bigquery
google-cloud-storage
pandas
PyYAML
